+++ apps/prairielearn/src/api/v1/endpoints/queries.sql
@@ -140,7 +140,6 @@ WITH
       aar.mode,
       aar.number AS assessment_access_rule_number,
       aar.password,
-      aar.seb_config,
       aar.show_closed_assessment,
       aar.show_closed_assessment_score,
       format_date_iso8601 (aar.start_date, ci.display_timezone) AS start_date,