+++ apps/prairielearn/src/pages/authLoginOAuth2/authLoginOAuth2.js
@@ -1,43 +1,38 @@
 // @ts-check
-const ERR = require('async-stacktrace');
-const express = require('express');
-const { OAuth2Client } = require('google-auth-library');
-const { logger } = require('@prairielearn/logger');
-const error = require('@prairielearn/error');
+const asyncHandler = require('express-async-handler');
+import { Router } from 'express';
+import { OAuth2Client } from 'google-auth-library';
+import { HttpStatusError } from '@prairielearn/error';
 
-const { config } = require('../../lib/config');
+import { config } from '../../lib/config';
 
-const router = express.Router();
+const router = Router();
 
-router.get('/', function (req, res, next) {
-  if (
-    !config.hasOauth ||
-    !config.googleClientId ||
-    !config.googleClientSecret ||
-    !config.googleRedirectUrl
-  ) {
-    return next(new error.HttpStatusError(404, 'Google login is not enabled'));
-  }
+router.get(
+  '/',
+  asyncHandler(async (req, res) => {
+    if (
+      !config.hasOauth ||
+      !config.googleClientId ||
+      !config.googleClientSecret ||
+      !config.googleRedirectUrl
+    ) {
+      throw new HttpStatusError(404, 'Google login is not enabled');
+    }
 
-  let url;
-  try {
     const oauth2Client = new OAuth2Client(
       config.googleClientId,
       config.googleClientSecret,
       config.googleRedirectUrl,
     );
-    url = oauth2Client.generateAuthUrl({
+    const url = oauth2Client.generateAuthUrl({
       access_type: 'online',
       scope: ['openid', 'profile', 'email'],
       prompt: 'select_account',
       // FIXME: should add some state here to avoid CSRF
     });
-  } catch (err) {
-    ERR(err, next);
-    return;
-  }
-  logger.verbose('Google auth URL redirect: ' + url);
-  res.redirect(url);
-});
+    res.redirect(url);
+  }),
+);
 
-module.exports = router;
+export default router;